synthetic/rabbitmq/requirements.txt
htcondor
wipac-dev-tools
wipac-rest-tools>=1.3.19
wipac-mqclient[rabbitmq]
